Display balance, income and expense info for the events

+++ resources/migrations/20170213211550-add-events-table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE events
 (id VARCHAR(30) PRIMARY KEY,
  user_id INTEGER REFERENCES users (id),
  transaction_date VARCHAR(10),
- amount NUMERIC,
+ amount INTEGER,
  recipient VARCHAR(50),
  type VARCHAR(30),
  created TIMESTAMP DEFAULT current_timestamp NOT NULL);
